// dashboard/src/lib/types.ts
export interface NutritionFacts {
  id: string
  calories: number
  fats: number
  protein: number
  carbs: number
  sugar: number
  cholesterol: number
  vitamin_a: number
  vitamin_c: number
  vitamin_d: number
  vitamin_e: number
  vitamin_k: number
  vitamin_b1: number
  vitamin_b2: number
  vitamin_b3: number
  vitamin_b5: number
  vitamin_b6: number
  vitamin_b7: number
  vitamin_b9: number
  vitamin_b12: number
  potassium: number
  calcium: number
  iron: number
  sodium: number
  timestamp: string
  imgURL: string
} 